Add new class Main for driver function, rename files to match bot's name
